refactor(customers): rename domains to companies and org list to accounts
