+++ apps/customers/src/routes/companies/$domain/components/CompanyInfoCard/CompanyInfoCard.tsx
