+++ apps/customers/src/routes/companies/$domain/components/CompanyInfoCard/index.ts
