+++ apps/customers/src/routes/companies/$domain/components/DomainUsersTable/components/UserResearchButton/UserResearchButton.tsx
