+++ apps/customers/src/routes/companies/$domain/components/DomainUsersTable/components/UserResearchButton/index.ts
