+++ apps/customers/src/routes/companies/components/DomainsTable/index.ts
