+++ apps/customers/src/routes/domains/page.tsx
@@ -1,286 +1,8 @@
-import type { RouterInputs } from "@superset/trpc";
-import { Button } from "@superset/ui/button";
-import { Input } from "@superset/ui/input";
-import { Label } from "@superset/ui/label";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@superset/ui/select";
-import { toast } from "@superset/ui/sonner";
-import { Switch } from "@superset/ui/switch";
-import { Tabs, TabsList, TabsTrigger } from "@superset/ui/tabs";
-import {
-	keepPreviousData,
-	useMutation,
-	useQuery,
-	useQueryClient,
-} from "@tanstack/react-query";
-import { createFileRoute } from "@tanstack/react-router";
-import { useEffect, useState } from "react";
-import { LuSearch } from "react-icons/lu";
+import { createFileRoute, redirect } from "@tanstack/react-router";
 
-import { RefreshButton } from "@/components/RefreshButton";
-import { SnapshotNote } from "@/components/SnapshotNote";
-import { useTRPC } from "@/trpc/react";
-
-import { DomainsTable } from "./components/DomainsTable";
-
-type DomainRollupInput = RouterInputs["customers"]["domainRollup"];
-
-const PAGE_SIZE = 50;
-
+/** Legacy URL — the domain rollup is now the Companies page. */
 export const Route = createFileRoute("/domains/")({
-	component: DomainsPage,
+	beforeLoad: () => {
+		throw redirect({ to: "/companies" });
+	},
 });
-
-function DomainsPage() {
-	const trpc = useTRPC();
-
-	const [view, setView] = useState<"all" | "pinned">("all");
-	const [search, setSearch] = useState("");
-	const [debouncedSearch, setDebouncedSearch] = useState("");
-	const [minUsers, setMinUsers] = useState(2);
-	const [includeFreemail, setIncludeFreemail] = useState(false);
-	const [health, setHealth] = useState<DomainRollupInput["health"]>("all");
-	const [trend, setTrend] = useState<DomainRollupInput["trend"]>("all");
-	const [sort, setSort] = useState<DomainRollupInput["sort"]>("users");
-	const [page, setPage] = useState(1);
-
-	useEffect(() => {
-		const timeout = setTimeout(() => setDebouncedSearch(search), 300);
-		return () => clearTimeout(timeout);
-	}, [search]);
-
-	// biome-ignore lint/correctness/useExhaustiveDependencies: reset pagination whenever a filter changes
-	useEffect(() => {
-		setPage(1);
-	}, [debouncedSearch, minUsers, includeFreemail, health, trend, sort]);
-
-	const { data, isLoading, isFetching, error } = useQuery(
-		trpc.customers.domainRollup.queryOptions(
-			{
-				page,
-				pageSize: PAGE_SIZE,
-				search: debouncedSearch || undefined,
-				minUsers,
-				includeFreemail,
-				health,
-				trend,
-				sort,
-			},
-			{ placeholderData: keepPreviousData },
-		),
-	);
-
-	const totalPages = data ? Math.max(1, Math.ceil(data.total / PAGE_SIZE)) : 1;
-
-	const queryClient = useQueryClient();
-	const pinned = useQuery(trpc.customers.pinnedDomains.queryOptions());
-	const pinnedSet = new Set(pinned.data?.rows.map((row) => row.domain) ?? []);
-	const togglePin = useMutation(
-		trpc.customers.setDomainPinned.mutationOptions({
-			onSuccess: (result, variables) => {
-				queryClient.invalidateQueries({
-					queryKey: trpc.customers.pinnedDomains.queryKey(),
-				});
-				toast.success(
-					result.pinned
-						? `Pinned ${variables.domain}`
-						: `Unpinned ${variables.domain}`,
-				);
-			},
-			onError: (mutationError) =>
-				toast.error(`Failed to update pin: ${mutationError.message}`),
-		}),
-	);
-	const handleTogglePin = (domain: string, shouldPin: boolean) =>
-		togglePin.mutate({ domain, pinned: shouldPin });
-
-	return (
-		<div className="space-y-6">
-			<div className="flex items-end justify-between">
-				<div>
-					<h1 className="text-3xl font-bold tracking-tight">Domains</h1>
-					<p className="text-muted-foreground">
-						Users grouped by email domain — spot multi-user companies without a
-						formal org
-					</p>
-				</div>
-				<div className="flex items-center gap-3">
-					<SnapshotNote snapshotAt={data?.snapshotAt} />
-					<RefreshButton />
-				</div>
-			</div>
-
-			<Tabs
-				value={view}
-				onValueChange={(value) => setView(value as "all" | "pinned")}
-			>
-				<TabsList>
-					<TabsTrigger value="all">All domains</TabsTrigger>
-					<TabsTrigger value="pinned">
-						Pinned{" "}
-						{pinnedSet.size > 0 && (
-							<span className="text-muted-foreground">({pinnedSet.size})</span>
-						)}
-					</TabsTrigger>
-				</TabsList>
-			</Tabs>
-
-			{view === "all" && (
-				<div className="flex flex-wrap items-center gap-6">
-					<div className="relative">
-						<LuSearch className="text-muted-foreground absolute top-1/2 left-2.5 size-4 -translate-y-1/2" />
-						<Input
-							value={search}
-							onChange={(event) => setSearch(event.target.value)}
-							placeholder="Search domains"
-							className="w-56 pl-8"
-						/>
-					</div>
-
-					<div className="flex items-center gap-2">
-						<Label
-							htmlFor="min-users"
-							className="text-muted-foreground text-sm"
-						>
-							Min users
-						</Label>
-						<Select
-							value={String(minUsers)}
-							onValueChange={(value) => setMinUsers(Number(value))}
-						>
-							<SelectTrigger id="min-users" className="w-20">
-								<SelectValue />
-							</SelectTrigger>
-							<SelectContent>
-								{[1, 2, 3, 5, 10].map((count) => (
-									<SelectItem key={count} value={String(count)}>
-										{count}+
-									</SelectItem>
-								))}
-							</SelectContent>
-						</Select>
-					</div>
-
-					<Select
-						value={health}
-						onValueChange={(value) =>
-							setHealth(value as NonNullable<DomainRollupInput["health"]>)
-						}
-					>
-						<SelectTrigger className="w-36">
-							<SelectValue placeholder="Health" />
-						</SelectTrigger>
-						<SelectContent>
-							<SelectItem value="all">All health</SelectItem>
-							<SelectItem value="active">Active</SelectItem>
-							<SelectItem value="idle">Idle</SelectItem>
-							<SelectItem value="cooling">Cooling</SelectItem>
-							<SelectItem value="dormant">Dormant</SelectItem>
-							<SelectItem value="churnRisk">Churn risk</SelectItem>
-						</SelectContent>
-					</Select>
-
-					<Select
-						value={trend}
-						onValueChange={(value) =>
-							setTrend(value as NonNullable<DomainRollupInput["trend"]>)
-						}
-					>
-						<SelectTrigger className="w-36">
-							<SelectValue placeholder="Trend" />
-						</SelectTrigger>
-						<SelectContent>
-							<SelectItem value="all">All trends</SelectItem>
-							<SelectItem value="growing">Growing</SelectItem>
-							<SelectItem value="declining">Declining</SelectItem>
-						</SelectContent>
-					</Select>
-
-					<Select
-						value={sort}
-						onValueChange={(value) =>
-							setSort(value as NonNullable<DomainRollupInput["sort"]>)
-						}
-					>
-						<SelectTrigger className="w-44">
-							<SelectValue placeholder="Sort" />
-						</SelectTrigger>
-						<SelectContent>
-							<SelectItem value="users">Most users</SelectItem>
-							<SelectItem value="events30d">Most events (30d)</SelectItem>
-							<SelectItem value="lastActive">Last active</SelectItem>
-							<SelectItem value="trend">Best trend</SelectItem>
-						</SelectContent>
-					</Select>
-
-					<div className="flex items-center gap-2">
-						<Switch
-							id="include-freemail"
-							checked={includeFreemail}
-							onCheckedChange={setIncludeFreemail}
-						/>
-						<Label
-							htmlFor="include-freemail"
-							className="text-muted-foreground text-sm"
-						>
-							Include freemail (gmail, outlook, …)
-						</Label>
-					</div>
-				</div>
-			)}
-
-			{view === "pinned" ? (
-				<DomainsTable
-					rows={pinned.data?.rows}
-					total={pinned.data?.rows.length}
-					isLoading={pinned.isLoading}
-					isFetching={pinned.isFetching}
-					error={null}
-					pinnedDomains={pinnedSet}
-					onTogglePin={handleTogglePin}
-					emptyTitle="No pinned domains"
-					emptyHint="Star a domain in the list or on its page to bookmark it here"
-				/>
-			) : (
-				<DomainsTable
-					rows={data?.rows}
-					total={data?.total}
-					isLoading={isLoading}
-					isFetching={isFetching}
-					error={error}
-					pinnedDomains={pinnedSet}
-					onTogglePin={handleTogglePin}
-				/>
-			)}
-
-			{view === "all" && data && data.total > PAGE_SIZE && (
-				<div className="flex items-center justify-end gap-3">
-					<span className="text-muted-foreground text-sm">
-						Page {page} of {totalPages}
-					</span>
-					<Button
-						variant="outline"
-						size="sm"
-						disabled={page <= 1}
-						onClick={() => setPage((current) => current - 1)}
-					>
-						Previous
-					</Button>
-					<Button
-						variant="outline"
-						size="sm"
-						disabled={page >= totalPages}
-						onClick={() => setPage((current) => current + 1)}
-					>
-						Next
-					</Button>
-				</div>
-			)}
-		</div>
-	);
-}
